Fractal Tree maker
Uses a method that I saw on coding train to add new branches to a branch object. It uses POO, what I'm trying to learn, and the branches can be manipulated in all kinds of ways. In the example on the code, it makes all branhches green after some point.

It was way harder than I thought because turtle does not calculate the end of a line based on angle and size in a staightforward way, so I created my own line object that is cabable of calculate rotations and size changes based on another turtle.

The turtle that draws is named bob, and the one who calculates is called roberto.